requirements.txt
# Web Framework & Dashboard
Flask
streamlit

# Computer Vision & Arrays
opencv-python
numpy
pandas
Pillow

# Biometric AI & Dependencies
cmake
dlib
face_recognition

# CRITICAL FIX: Downgrade setuptools to prevent pkg_resources module crash in face_recognition_models
setuptools<70.0.0